style: 반응형 css 구현 #164
